Fix cypress config eslint problems
