+++ cypress/support/e2e.ts
@@ -15,20 +15,4 @@
 
 import '@cypress/code-coverage/support';
 import 'cypress-axe';
-import { Notification } from '../../src/types/notification.types';
-import { RoleName } from '../../src/types/user.types';
 import './commands';
-
-declare global {
-  namespace Cypress {
-    interface Chainable {
-      mocklogin(): void;
-      setUserRolesInRedux(roles: RoleName[]): void;
-      startRegistrationWithDoi(): void;
-      selectNpiDiscipline(discipline: string): void;
-      setNotificationInRedux(notification: Notification): void;
-      removeNotificationInRedux(): void;
-      mockFileUpload(): void;
-    }
-  }
-}
